# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Spin API service
  description: HTTP service for controlling the spin service
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /storage/images/create:
    post:
      tags:
      - spin-registry
      summary: storage_images_create spin-registry
      description: add an image definition to the registry
      operationId: spin-registry#storage_images_create
      parameters:
      - name: storage_images_create_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageImagesCreateRequestBody'
          required:
          - image
          - volume
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/SpinRegistryStorageImagesCreateResponseBody'
            required:
            - path
            - cdrom
      schemes:
      - http
  /storage/images/delete:
    post:
      tags:
      - spin-registry
      summary: storage_images_delete spin-registry
      description: remove an image definition from the registry
      operationId: spin-registry#storage_images_delete
      parameters:
      - name: storage_images_delete_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageImagesDeleteRequestBody'
          required:
          - volume_name
          - image_name
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /storage/images/get:
    get:
      tags:
      - spin-registry
      summary: storage_images_get spin-registry
      description: retrieves an image definition from the registry
      operationId: spin-registry#storage_images_get
      parameters:
      - name: storage_images_get_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageImagesGetRequestBody'
          required:
          - volume_name
          - image_name
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/SpinRegistryStorageImagesGetResponseBody'
            required:
            - path
            - cdrom
      schemes:
      - http
  /storage/images/list:
    get:
      tags:
      - spin-registry
      summary: storage_images_list spin-registry
      description: list all images by volume
      operationId: spin-registry#storage_images_list
      parameters:
      - name: storage_images_list_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageImagesListRequestBody'
          required:
          - volume_name
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              type: string
              example: Vel in modi rerum.
      schemes:
      - http
  /storage/volumes/create:
    post:
      tags:
      - spin-registry
      summary: storage_volumes_create spin-registry
      description: create a new volume
      operationId: spin-registry#storage_volumes_create
      parameters:
      - name: storage_volumes_create_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageVolumesCreateRequestBody'
          required:
          - name
          - path
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /storage/volumes/delete:
    post:
      tags:
      - spin-registry
      summary: storage_volumes_delete spin-registry
      description: delete an existing volume
      operationId: spin-registry#storage_volumes_delete
      parameters:
      - name: storage_volumes_delete_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryStorageVolumesDeleteRequestBody'
          required:
          - name
      responses:
        "200":
          description: OK response.
      schemes:
      - http
...
  /vm/create:
    post:
      tags:
      - spin-registry
      summary: vm_create spin-registry
      description: Create a VM
      operationId: spin-registry#vm_create
      parameters:
      - name: vm_create_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryVMCreateRequestBody'
          required:
          - name
          - cpus
          - memory
          - images
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /vm/delete/{id}:
    post:
      tags:
      - spin-registry
      summary: vm_delete spin-registry
      description: Delete a VM by ID
      operationId: spin-registry#vm_delete
      parameters:
      - name: id
        in: path
        description: ID of VM to remove
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
      schemes:
      - http
...
  /vm/update/{id}:
    post:
      tags:
      - spin-registry
      summary: vm_update spin-registry
      description: Update a VM
      operationId: spin-registry#vm_update
      parameters:
      - name: id
        in: path
        description: ID of VM to update
        required: true
        type: integer
      - name: vm_update_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryVMUpdateRequestBody'
          required:
          - vm
      responses:
        "200":
          description: OK response.
      schemes:
      - http
definitions:
  ImageRequestBody:
    title: ImageRequestBody
    type: object
    properties:
      cdrom:
        type: boolean
        description: Is this a cdrom image?
        example: false
      path:
        type: string
        description: Image path
        example: Sed necessitatibus et asperiores.
    example:
      cdrom: false
      path: Sequi officia animi fugiat.
    required:
    - path
    - cdrom
...
  SpinRegistryStorageImagesCreateRequestBody:
    title: SpinRegistryStorageImagesCreateRequestBody
    type: object
    properties:
      cdrom:
        type: boolean
        description: Is this image a cdrom?
        example: true
      image:
        type: string
        description: Image filename, no `/` characters
        example: Animi dolorem totam omnis molestiae voluptatum rerum.
      image_size:
        type: integer
        description: Image size (in gigabytes)
        example: 6630383005671600344
        format: int64
      volume:
        type: string
        description: Volume name
        example: Optio sint et nesciunt et molestiae soluta.
    example:
      cdrom: false
      image: Deserunt blanditiis consequatur consequuntur praesentium consequatur.
      image_size: 710111403244910451
      volume: Tempore iste voluptatem est.
    required:
    - image
    - volume
  SpinRegistryStorageImagesCreateResponseBody:
    title: SpinRegistryStorageImagesCreateResponseBody
    type: object
    properties:
      cdrom:
        type: boolean
        description: Is this a cdrom image?
        example: true
      path:
        type: string
        description: Image path
        example: Asperiores tempora.
    example:
      cdrom: false
      path: Inventore qui nihil nam.
    required:
    - path
    - cdrom
  SpinRegistryStorageImagesDeleteRequestBody:
    title: SpinRegistryStorageImagesDeleteRequestBody
    type: object
    properties:
      image_name:
        type: string
        description: name of image
        example: Labore numquam corporis ab nulla.
      volume_name:
        type: string
        description: name of volume
        example: Harum rerum.
    example:
      image_name: Minus expedita.
      volume_name: Exercitationem voluptatem vel dicta consequuntur tempora dolorem.
    required:
    - volume_name
    - image_name
  SpinRegistryStorageImagesGetRequestBody:
    title: SpinRegistryStorageImagesGetRequestBody
    type: object
    properties:
      image_name:
        type: string
        description: name of image
        example: Culpa ut quidem aut et qui porro.
      volume_name:
        type: string
        description: name of volume
        example: Minima omnis.
    example:
      image_name: Totam quia sit.
      volume_name: Labore accusantium ut itaque odio explicabo perferendis.
    required:
    - volume_name
    - image_name
  SpinRegistryStorageImagesGetResponseBody:
    title: SpinRegistryStorageImagesGetResponseBody
    type: object
    properties:
      cdrom:
        type: boolean
        description: Is this a cdrom image?
        example: true
      path:
        type: string
        description: Image path
        example: Et iusto fugit.
    example:
      cdrom: true
      path: Aliquid corrupti amet numquam sed.
    required:
    - path
    - cdrom
  SpinRegistryStorageImagesListRequestBody:
    title: SpinRegistryStorageImagesListRequestBody
    type: object
    properties:
      volume_name:
        type: string
        description: name of volume to list images for
        example: Ut veritatis aut.
    example:
      volume_name: Deleniti incidunt.
    required:
    - volume_name
  SpinRegistryStorageVolumesCreateRequestBody:
    title: SpinRegistryStorageVolumesCreateRequestBody
    type: object
    properties:
      name:
        type: string
        description: name of volume
        example: Inventore aut consequatur officiis perspiciatis inventore.
      path:
        type: string
        description: path to volume
        example: Voluptate voluptas veniam.
    example:
      name: Eum laudantium tempora pariatur possimus autem.
      path: Qui ad aut possimus quisquam harum.
    required:
    - name
    - path
  SpinRegistryStorageVolumesDeleteRequestBody:
    title: SpinRegistryStorageVolumesDeleteRequestBody
    type: object
    properties:
      name:
        type: string
        description: name of volume
        example: Ipsum natus quasi quia.
    example:
      name: Ut suscipit cupiditate cumque ratione occaecati.
    required:
    - name
  SpinRegistryVMCreateRequestBody:
    title: SpinRegistryVMCreateRequestBody
    type: object
    properties:
      cpus:
        type: integer
        description: CPU count
        example: 743273099863973551
        format: int64
      images:
        type: array
        items:
          $ref: '#/definitions/ImageRequestBody'
        description: Image references
        example:
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
      memory:
        type: integer
        description: Memory (in megabytes)
        example: 12485697019927615723
        format: int64
      name:
        type: string
        description: Name of VM; does not need to be unique
        example: Aut tempore aliquam.
    example:
      cpus: 848682044187229832
      images:
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      memory: 865562045241744601
      name: Natus neque qui iusto ex et.
    required:
    - name
    - cpus
    - memory
    - images
...
  SpinRegistryVMUpdateRequestBody:
    title: SpinRegistryVMUpdateRequestBody
    type: object
    properties:
      vm:
        $ref: '#/definitions/UpdatedVMRequestBody'
    example:
      vm:
        cpus: 17043342971635177953
        images:
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        memory: 2551606264180670795
        name: Rem et nobis.
    required:
    - vm
...
  UpdatedVMRequestBody:
    title: UpdatedVMRequestBody
    type: object
    properties:
      cpus:
        type: integer
        description: CPU count
        example: 12912299026057686733
        format: int64
      images:
        type: array
        items:
          $ref: '#/definitions/ImageRequestBody'
        description: Image references
        example:
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
      memory:
        type: integer
        description: Memory (in megabytes)
        example: 3217507219102201512
        format: int64
      name:
        type: string
        description: Name of VM; does not need to be unique
        example: Reiciendis eveniet repudiandae reiciendis perferendis impedit reiciendis.
    example:
      cpus: 16407373559890724020
      images:
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      memory: 9724042061409375964
      name: Iure nihil.
    required:
    - name
    - cpus
    - memory
    - images